refined

+++ src/componenets/reservations/ReservationList.jsx
@@ -1,26 +1,55 @@
 import { useState, useEffect } from 'react'
+import axios from 'axios'
 import '../../ReservationList.css'
-import { Link, useParams, useNavigate } from "react-router-dom";
+import { Link, useParams } from "react-router-dom";
 
 export default function () {
     const [data, setData] = useState([])
     const { id } = useParams()
-    const redirectMe = useNavigate();
 
     useEffect(() => {
-        function fetchData() {
-            fetch('http://localhost:8000/reservations')
-                .then((response) => response.json())
-                .then((data) => setData(data))
-                .catch((error) => {
-                    console.clear()
-                    console.error('Error fetching data:', error)
-                });
+        // function fetchData() {
+        //     fetch('http://localhost:8000/reservations')
+        //         .then((response) => response.json())
+        //         .then((data) => setData(data))
+        //         .catch((error) => {
+        //             console.clear()
+        //             console.error('Error fetching data:', error)
+        //         });
 
-            // async / await (try catch)
+        //     // async / await (try catch)
+        // }
+
+        // fetchData();
+
+        // const getData = async () => {
+        //     try {
+        //         const response = await fetch("http://localhost:8000/reservations");
+
+        //         if (response.ok) {
+        //             const data = await response.json();
+        //             setData(data);
+        //             console.log(data);
+        //         }else
+        //             throw new Error("Erreur HTTP");
+
+        //     } catch (error) {
+        //         console.error(error);
+        //     }
+        // }
+
+        // getData();
+
+        async function getData() {
+            try {
+                const response = await axios.get("http://localhost:8000/reservations");
+                let newData = response.data;
+                setData(newData);
+            } catch (error) {
+                console.error(error);
+            }
         }
-
-        fetchData();
+        getData();
 
         return () => {
             console.log("Cleanup if needed");
@@ -29,31 +58,31 @@ export default function () {
 
     // cycle de vie: Montage => Mise à jour => Démontage
     const deleteObject = (targetId = null) => {
-        // let confirmation = confirm("Voulez-vous vraiment supprimer cette réservation ?")
-        // if (confirmation && targetId) {
-        // console.log("ID delete:", typeof targetId + " => " + targetId);
-        function deleteData() {
-            fetch(`http://localhost:8000/reservations/${targetId}`, {
-                method: 'DELETE' // verbe HTTP
-            })
-                .then((response) => response.json())
-                .then((dataAfterDelete) => {
-                    setData(
-                        data.filter(
-                            (element) => {
-                                return element.id != targetId
-                            }
-                        )
-                    )
+        let confirmation = confirm("Voulez-vous vraiment supprimer cette réservation ?")
+        if (confirmation && targetId) {
+            console.log("ID delete:", typeof targetId + " => " + targetId);
+            function deleteData() {
+                fetch(`http://localhost:8000/reservations/${targetId}`, {
+                    method: 'DELETE' // verbe HTTP
                 })
-                .catch((error) => {
-                    console.clear()
-                    console.error('Error de suppression:', error)
-                });
-        }
+                    .then((response) => response.json())
+                    .then((dataAfterDelete) => {
+                        setData(
+                            data.filter(
+                                (element) => {
+                                    return element.id != targetId
+                                }
+                            )
+                        )
+                    })
+                    .catch((error) => {
+                        console.clear()
+                        console.error('Error de suppression:', error)
+                    });
+            }
 
-        deleteData();
-        // }
+            deleteData();
+        }
     }
 
 
@@ -61,7 +90,7 @@ export default function () {
         <>
             <div className="reservation-list">
                 <div>
-                    <Link to="/add-reservation" style={{ color: '#fff', textDecoration: 'none', fontSize: '16px' }}>Ajouter</Link>
+                    <Link to="/add-reservation">Ajouter</Link>
                 </div>
                 {
                     data.map(
